+++ frame/supports/support-command-kv/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "support-command-kv"
+version = "0.5.0"
+authors = ["Darwinia Network <hello@darwinia.network>"]
+description = "Darwinia bridger"
+repository = "https://github.com/darwinia-network/bridger"
+license = "MIT"
+documentation = "https://rust-docs.darwinia.network/bridger"
+homepage = "https://github.com/darwinia-network/bridger"
+include = [
+  "Cargo.toml",
+  "**/*.rs",
+  "README.md",
+  "LICENSE"
+]
+keywords = ["substrate", "darwinia"]
+readme = 'README.md'
+edition = "2021"
+
+[dependencies]
+
+structopt = "0.3"
+
+tracing = "0.1"
+color-eyre = "0.5"
+term-table = "1.3"
+
+support-common     = { path = "../support-common" }
+support-terminal   = { path = "../support-terminal" }
+
+component-state    = { path = "../../components/state" }
